Add all simple requests with macro

+++ src/x11/requests/create_window.rs
@@ -15,7 +15,7 @@ impl Request for CreateWindow {
 		1
 	}
 
-	fn minor_opcode() -> Option<u16> {
+	fn minor_opcode() -> Option<u8> {
 		// This is an X core protocol request: it is not an extension, and so
 		// therefore does not use the `minor_opcode`.
 		None
